requirements.txt
flask
modbus_tk
pyserial
